+++ projects/vdk-core/tests/functional/run/test_run_sql_queries.py
@@ -8,19 +8,67 @@ from unittest import mock
 
 from click.testing import Result
 from functional.run import util
+from vdk.api.plugin.hook_markers import hookimpl
+from vdk.internal.builtin_plugins.connection.pep249.interfaces import PEP249Connection
+from vdk.internal.builtin_plugins.connection.recovery_cursor import RecoveryCursor
+from vdk.internal.builtin_plugins.run.job_context import JobContext
 from vdk.internal.core.errors import VdkConfigurationError
 from vdk.plugin.test_utils.util_funcs import cli_assert_equal
 from vdk.plugin.test_utils.util_funcs import CliEntryBasedTestRunner
 from vdk.plugin.test_utils.util_funcs import get_test_job_path
 from vdk.plugin.test_utils.util_plugins import DB_TYPE_SQLITE_MEMORY
 from vdk.plugin.test_utils.util_plugins import DecoratedSqLite3MemoryDbPlugin
+from vdk.plugin.test_utils.util_plugins import SqLite3MemoryConnection
 from vdk.plugin.test_utils.util_plugins import SqLite3MemoryDbPlugin
-from vdk.plugin.test_utils.util_plugins import SyntaxErrorRecoverySqLite3MemoryDbPlugin
-from vdk.plugin.test_utils.util_plugins import ValidatedSqLite3MemoryDbPlugin
 
 log = logging.getLogger(__name__)
 
 VDK_DB_DEFAULT_TYPE = "VDK_DB_DEFAULT_TYPE"
+
+
+class ValidatedSqLite3MemoryDbPlugin:
+    def new_connection(self) -> PEP249Connection:
+        return SqLite3MemoryConnection()
+
+    @hookimpl
+    def initialize_job(self, context: JobContext) -> None:
+        context.connections.add_open_connection_factory_method(
+            DB_TYPE_SQLITE_MEMORY, self.new_connection
+        )
+
+    @hookimpl(trylast=True)
+    def db_connection_validate_operation(self, operation, parameters):
+        parameters_length = 0
+        if parameters:
+            parameters_length = len("".join(map(str, parameters)))
+
+        if len(operation) + parameters_length > 10000:
+            raise Exception(
+                "Database operation has exceeded the maximum limit of 10000 characters."
+            )
+
+
+class SyntaxErrorRecoverySqLite3MemoryDbPlugin:
+    def __init__(self):
+        self._max_retries = 5
+
+    def new_connection(self) -> PEP249Connection:
+        return SqLite3MemoryConnection()
+
+    @hookimpl
+    def initialize_job(self, context: JobContext) -> None:
+        context.connections.add_open_connection_factory_method(
+            DB_TYPE_SQLITE_MEMORY, self.new_connection
+        )
+
+    @hookimpl(trylast=True)
+    def db_connection_recover_operation(self, recovery_cursor: RecoveryCursor) -> None:
+        managed_operation = recovery_cursor.get_managed_operation()
+        if "syntax error".upper() in recovery_cursor.get_exception().args[0].upper():
+            recovery_cursor.execute(
+                managed_operation.get_operation().replace("Syntax error", "123.0")
+            )
+            recovery_cursor.retries_increment()
 
 
 @mock.patch.dict(os.environ, {VDK_DB_DEFAULT_TYPE: DB_TYPE_SQLITE_MEMORY})
