+++ projects/vdk-plugins/vdk-test-utils/src/vdk/plugin/test_utils/util_plugins.py
@@ -18,7 +18,6 @@ from vdk.internal.builtin_plugins.connection.managed_connection_base import (
     ManagedConnectionBase,
 )
 from vdk.internal.builtin_plugins.connection.pep249.interfaces import PEP249Connection
-from vdk.internal.builtin_plugins.connection.recovery_cursor import RecoveryCursor
 from vdk.internal.builtin_plugins.run.job_context import JobContext
 from vdk.internal.util.decorators import closing_noexcept_on_close
 
@@ -84,51 +83,6 @@ class SqLite3MemoryConnection(ManagedConnectionBase):
 
     def _connect(self) -> PEP249Connection:
         return self.db.new_connection()
-
-
-class ValidatedSqLite3MemoryDbPlugin:
-    def new_connection(self) -> PEP249Connection:
-        return SqLite3MemoryConnection()
-
-    @hookimpl
-    def initialize_job(self, context: JobContext) -> None:
-        context.connections.add_open_connection_factory_method(
-            DB_TYPE_SQLITE_MEMORY, self.new_connection
-        )
-
-    @hookimpl(trylast=True)
-    def db_connection_validate_operation(self, operation, parameters):
-        parameters_length = 0
-        if parameters:
-            parameters_length = len("".join(map(str, parameters)))
-
-        if len(operation) + parameters_length > 10000:
-            raise Exception(
-                "Database operation has exceeded the maximum limit of 10000 characters."
-            )
-
-
-class SyntaxErrorRecoverySqLite3MemoryDbPlugin:
-    def __init__(self):
-        self._max_retries = 5
-
-    def new_connection(self) -> PEP249Connection:
-        return SqLite3MemoryConnection()
-
-    @hookimpl
-    def initialize_job(self, context: JobContext) -> None:
-        context.connections.add_open_connection_factory_method(
-            DB_TYPE_SQLITE_MEMORY, self.new_connection
-        )
-
-    @hookimpl(trylast=True)
-    def db_connection_recover_operation(self, recovery_cursor: RecoveryCursor) -> None:
-        managed_operation = recovery_cursor.get_managed_operation()
-        if "syntax error".upper() in recovery_cursor.get_exception().args[0].upper():
-            recovery_cursor.execute(
-                managed_operation.get_operation().replace("Syntax error", "123.0")
-            )
-            recovery_cursor.retries_increment()
 
 
 class DecoratedSqLite3MemoryDbPlugin:
